düzeltme

+++ src/main/java/com/turkcell/spring/workshop/repositories/OrderDetailsRepository.java
@@ -3,7 +3,7 @@ package com.turkcell.spring.workshop.repositories;
 import com.turkcell.spring.workshop.entities.OrderDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderDetailsRepository extends JpaRepository<OrderDetail, Integer> {
+public interface OrderDetailsRepository extends JpaRepository<OrderDetail, Short> {
 
 
 }
